+++ src/components/ui/Icon/Icon.test.tsx
@@ -2,7 +2,7 @@ import { cleanup, render, screen } from "@testing-library/react";
 import React from "react";
 import { afterEach, describe, expect, it } from "vitest";
 
-import { getColor } from "@/component/common";
+import { getColor } from "@/components/common";
 
 import { Icon } from "./Icon";
